Make Vector into a separate subproject
Create vector as a separate project. Thus instead of adding all targets
in the master CMakeLists.txt we can now create separate projects and add
them in master CMakeLists.txt with minimal changes.
Note that we've included the directories in the top level cmake to be
included for all the sub projects.

Signed-off-by: Pranay Kr. Srivastava <pranjas@gmail.com>